requirements.txt
nicegui>=2.0
yt-dlp
pyyaml
nicegui